+++ src/nidas/core/SampleIOProcessor.cc
@@ -169,7 +169,6 @@ void SampleIOProcessor::fromDOMElement(const xercesc::DOMElement* node)
 		}
 		setOptional(val);
 	    }
-            else if (aname == "xml:base" || aname == "xmlns") {}
         }
     }