+++ src/nidas/core/Site.cc
@@ -193,7 +193,6 @@ void Site::fromDOMElement(const xercesc::DOMElement* node)
                 }
                 _applyCals = val;
             }
-            else if (aname == "xml:base" || aname == "xmlns") {}
         }
     }